# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"MyService": {
				"properties": {
					"serviceId": {
						"type": "string"
					},
					"serviceName": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"category": {
						"$ref": "#/components/schemas/Category"
					}
				},
				"required": [
					"serviceId",
					"serviceName",
					"price",
					"isActive",
					"createdAt",
					"updatedAt",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"properties": {
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/MyService"
						},
						"type": "array"
					}
				},
				"required": [
					"categoryId",
					"categoryName",
					"isActive",
					"createdAt",
					"updatedAt",
					"services"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WorkPortfolio": {
				"properties": {
					"workId": {
						"type": "string"
					},
					"serviceId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"workDate": {
						"type": "string",
						"format": "date-time"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"service": {
						"$ref": "#/components/schemas/MyService"
					}
				},
				"required": [
					"workId",
					"serviceId",
					"title",
					"imageUrl",
					"workDate",
					"isActive",
					"createdAt",
					"updatedAt",
					"service"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginationWithData_WorkPortfolio_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/WorkPortfolio"
						},
						"type": "array"
					},
					"totalSize": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalSize",
					"pageSize",
					"pageNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MyWorkRequest": {
				"properties": {
					"serviceId": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pageSize",
					"pageNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Boolean": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"MyWorkCreate": {
				"properties": {
					"serviceId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"serviceId",
					"title",
					"description",
					"imageUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Express.Multer.File": {
				"description": "Object containing file metadata and access information.",
				"properties": {
					"fieldname": {
						"type": "string",
						"description": "Name of the form field associated with this file."
					},
					"originalname": {
						"type": "string",
						"description": "Name of the file on the uploader's computer."
					},
					"encoding": {
						"type": "string",
						"description": "Value of the `Content-Transfer-Encoding` header for this file.",
						"deprecated": true
					},
					"mimetype": {
						"type": "string",
						"description": "Value of the `Content-Type` header for this file."
					},
					"size": {
						"type": "number",
						"format": "double",
						"description": "Size of the file in bytes."
					},
					"stream": {
						"type": "string",
						"format": "byte",
						"description": "A readable stream of this file. Only available to the `_handleFile`\r\ncallback for custom `StorageEngine`s."
					},
					"destination": {
						"type": "string",
						"description": "`DiskStorage` only: Directory to which this file has been uploaded."
					},
					"filename": {
						"type": "string",
						"description": "`DiskStorage` only: Name of this file within `destination`."
					},
					"path": {
						"type": "string",
						"description": "`DiskStorage` only: Full path to the uploaded file."
					},
					"buffer": {
						"type": "string",
						"format": "byte",
						"description": "`MemoryStorage` only: A Buffer containing the entire file."
					}
				},
				"required": [
					"fieldname",
					"originalname",
					"encoding",
					"mimetype",
					"size",
					"stream",
					"destination",
					"filename",
					"path",
					"buffer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceCreate": {
				"properties": {
					"serviceId": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"serviceName": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"imageUrl": {
						"$ref": "#/components/schemas/Express.Multer.File"
					}
				},
				"required": [
					"categoryId",
					"serviceName",
					"price",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Client": {
				"properties": {
					"clientId": {
						"type": "string"
					},
					"clientName": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"gmail": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"clientId",
					"clientName",
					"phoneNumber",
					"gmail",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryCreate": {
				"properties": {
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"categoryName",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookingStatus": {
				"type": "string",
				"enum": [
					"OTP Pending",
					"Pending",
					"Confirmed",
					"Completed",
					"Cancelled"
				]
			},
			"Booking": {
				"properties": {
					"bookingId": {
						"type": "string"
					},
					"bookingNumber": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"bookingDate": {
						"type": "string"
					},
					"bookingTime": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/BookingStatus"
					},
					"notes": {
						"type": "string"
					},
					"reviewRating": {
						"type": "number",
						"format": "double"
					},
					"reviewText": {
						"type": "string"
					},
					"reviewDate": {
						"type": "string",
						"format": "date-time"
					},
					"isOtpVerified": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"client": {
						"$ref": "#/components/schemas/Client"
					},
					"bookingServices": {
						"items": {
							"$ref": "#/components/schemas/MyBookingServices"
						},
						"type": "array"
					}
				},
				"required": [
					"bookingId",
					"bookingNumber",
					"clientId",
					"bookingDate",
					"bookingTime",
					"location",
					"totalPrice",
					"status",
					"isOtpVerified",
					"isActive",
					"createdAt",
					"updatedAt",
					"client",
					"bookingServices"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MyBookingServices": {
				"properties": {
					"bookingServiceId": {
						"type": "string"
					},
					"bookingId": {
						"type": "string"
					},
					"serviceId": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"booking": {
						"$ref": "#/components/schemas/Booking"
					},
					"service": {
						"$ref": "#/components/schemas/MyService"
					}
				},
				"required": [
					"bookingServiceId",
					"bookingId",
					"serviceId",
					"price",
					"isActive",
					"createdAt",
					"updatedAt",
					"booking",
					"service"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBookingService": {
				"properties": {
					"bookingId": {
						"type": "string"
					},
					"serviceId": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"serviceId",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBooking": {
				"properties": {
					"bookedServices": {
						"items": {
							"$ref": "#/components/schemas/CreateBookingService"
						},
						"type": "array"
					},
					"bookingDate": {
						"type": "string"
					},
					"bookingTime": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"clientName": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"gmail": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"bookedServices",
					"bookingDate",
					"bookingTime",
					"location",
					"totalPrice",
					"clientName",
					"phoneNumber",
					"gmail",
					"notes"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateBookingReview": {
				"properties": {
					"bookingNumber": {
						"type": "string"
					},
					"clientNumber": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"review": {
						"type": "string"
					}
				},
				"required": [
					"bookingNumber",
					"clientNumber",
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClientBooking": {
				"properties": {
					"phoneNumber": {
						"type": "string"
					},
					"bookingNumber": {
						"type": "string"
					}
				},
				"required": [
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyBooking": {
				"properties": {
					"bookingNumber": {
						"type": "string"
					},
					"gmail": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"required": [
					"bookingNumber",
					"gmail",
					"otp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResendOtp": {
				"properties": {
					"bookingNumber": {
						"type": "string"
					},
					"gmail": {
						"type": "string"
					}
				},
				"required": [
					"bookingNumber",
					"gmail"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "glow_by_rishi_api",
		"version": "1.0.0",
		"description": "It is a website for a beautican",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/my-works/all": {
			"post": {
				"operationId": "GetMyWorks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginationWithData_WorkPortfolio_"
								}
							}
						}
					}
				},
				"tags": [
					"My Works"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MyWorkRequest"
							}
						}
					}
				}
			}
		},
		"/api/my-works": {
			"post": {
				"operationId": "CreateMyWork",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"My Works"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MyWorkCreate"
							}
						}
					}
				}
			}
		},
		"/api/my-works/{myWorkId}": {
			"put": {
				"operationId": "UpdatemyWork",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"My Works"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "myWorkId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MyWorkCreate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMyWork",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"My Works"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "myWorkId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/services/services-by-category/{categoryId}": {
			"get": {
				"operationId": "GetServicesByCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MyService"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"My Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/services": {
			"post": {
				"operationId": "CreateService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"My Services"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ServiceCreate"
							}
						}
					}
				}
			}
		},
		"/api/services/{serviceId}": {
			"put": {
				"operationId": "UpdateService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"My Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ServiceCreate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"My Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/categories/{categoryId}": {
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryCreate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/categories": {
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryCreate"
							}
						}
					}
				}
			}
		},
...
		"/api/bookings/create": {
			"post": {
				"operationId": "CreateBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Booking"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBooking"
							}
						}
					}
				}
			}
		},
...
		"/api/bookings": {
			"get": {
				"operationId": "GetAllBookingsByStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Booking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BookingStatus"
						}
					}
				]
			}
		},
		"/api/bookings/{bookingId}": {
			"get": {
				"operationId": "GetBookingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Booking"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/bookings/create-review": {
			"post": {
				"operationId": "CreateBookingReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookingReview"
							}
						}
					}
				}
			}
		},
		"/api/bookings/update": {
			"post": {
				"operationId": "UpdateBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Booking"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Booking"
							}
						}
					}
				}
			}
		},
		"/api/bookings/update-status": {
			"post": {
				"operationId": "UpdateBookingStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"$ref": "#/components/schemas/BookingStatus"
									},
									"bookingId": {
										"type": "string"
									}
								},
								"required": [
									"status",
									"bookingId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/bookings/client-bookings/{clientId}": {
			"get": {
				"operationId": "GetBookingsByClientId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Booking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/bookings/client-bookings": {
			"post": {
				"operationId": "GetClientBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Booking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClientBooking"
							}
						}
					}
				}
			}
		},
		"/api/bookings/client-bookings/{phoneNumber}": {
			"get": {
				"operationId": "GetBookingsByClientPhoneNumber",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Booking"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "phoneNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/bookings/verify-booking": {
			"post": {
				"operationId": "VerifyBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyBooking"
							}
						}
					}
				}
			}
		},
		"/api/bookings/resend-otp": {
			"post": {
				"operationId": "ResendOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResendOtp"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}